Add -opencl option
This adds an OpenCL implementation of the compressor.

Currently this can only do etc1_optimizer_compute() on a GPU. It still does the
CPU etc1_optimizer_compute() to check the results. In the future, GPU/CPU
checking would only be done if an extra flag was passed in.

This also compiles the OpenCL source at runtime - good for debug builds so
the error message shows the source line number. Release builds should pull
the OpenCL source into the binary, instead of looking for it in the filesystem.