Use stb_image instead of lodepng to load/save images
This means that we now support all formats supported by stb_image for
the purpose of transcoding. This is mostly valuable for JPEG data since
glTF content can have PNG or JPG images, and TGA data since this is a
common image interchange format for real-time graphics.

While it was technically possible to keep lodepng for saving images it
didn't seem like there's a strong reason to do so.

Note that stb libraries are compiled as part of stb_image.cpp to reduce
compile-time impact on basisu_enc.cpp.