Optimize ETC1 selector error computation
Using a similar observation to the previous change, we don't need to
evaluate up to 16 color distances for each of the 64 palette entries -
instead, we can precompute the per-pixel errors introduced by changing
the selector ahead of time.

This reduces the amount of color_distance evaluation which is pretty
expensive, and further improves the encode times - on a 2Kx2K image on a
Core i7-8700K, previous change reduced the time from 8.5 seconds to 7.3
seconds, and this one further reduces it to 6.6 seconds.