Precompute g_bc7_mode_N tables for solid color blocks
uastc_init takes 140ms in Chrome; this delays time-to-first-paint with
textures by 140ms per worker. The tables are just 2KB and 1KB
respectively; additionally, errors for mode 5 are all 0, and errors for
mode 6 are all 0 except for two values (0 and 255), so we can avoid
storing the errors and compute them from color values.

This reduces uastc_init cost to 30ms.